python_services/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
opencv-python==4.11.0.86
numpy==2.3.2
pillow==11.3.0
scipy==1.16.1
scikit-learn==1.7.1
requests==2.32.5
python-multipart==0.0.20